C:/Program Files/Git/anuncios/{id}/alterarstatus

+++ src/main/java/com/example/Iclean/services/AnuncioService.java
@@ -46,6 +46,7 @@ public class AnuncioService {
 	}
 
 	public void delete(Long id) {
+		authService.validateSelfOrAdmin(id);
 		try {
 			repository.deleteById(id);
 		} catch (EmptyResultDataAccessException e) {
